# api/openapi.yaml
openapi: 3.1.0
info:
  title: OReilly Customers Service API
  version: 1.0.0
  description: |
    Clean API-first customer ownership contract for the workshop pilot.
  x-owner: customer-platform
  x-domain: customers
  x-workshop-pattern: live-entity-owner
  x-api-lifecycle: workshop-pilot
servers:
  - url: https://api.example.com/customers
tags:
  - name: customers
    description: Live customer records owned by the customer domain.
paths:
  /customers/ar/{customerNumber}:
    get:
      tags:
        - customers
      summary: Get an AR customer
      operationId: getArCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerNumber'
      responses:
        '200':
          description: AR customer record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArCustomer'
              examples:
                default:
                  value:
                    customerNumber: 90001234
                    customerName: OReilly Test Garage
                    customerAddress: 123 Pilot Way
                    customerCity: Springfield
                    customerState: MO
                    customerZipCode: '65807'
                    customerPhone: 417-555-0188
                    activityCode: ACTIVE
        '404':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/NotFound'
  /customers/pro/{customerNumber}:
    get:
      tags:
        - customers
      summary: Get a Pro customer
      operationId: getProCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerNumber'
      responses:
        '200':
          description: Pro customer record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProCustomer'
              examples:
                default:
                  value:
                    customerType: pro
                    customerNumber: 90001234
                    customerName: OReilly Test Garage
                    customerAddress: 123 Pilot Way
                    customerCity: Springfield
                    customerState: MO
                    customerZipCode: '65807'
                    customerPhone: 417-555-0188
                    accountNotes: Preferred commercial account
                    activityCode: ACTIVE
                    customerTypeDesc: Professional installer
        '404':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/NotFound'
  /customers/cash/{customerNumber}:
    get:
      tags:
        - customers
      summary: Get a Cash customer
      operationId: getCashCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerNumber'
      responses:
        '200':
          description: Cash customer record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashCustomer'
              examples:
                default:
                  value:
                    customerType: cash
                    customerNumbers:
                      - 70004567
                    firstName: Casey
                    lastName: Driver
                    preferredPhone: 417-555-0177
                    preferredEmail: casey.driver@example.com
                    addresses:
                      - line1: 55 Market Street
                        city: Springfield
                        state: MO
                        postalCode: '65806'
                        countryCode: US
        '404':
          $ref: '../shared-contracts/common-schemas.yaml#/components/responses/NotFound'
components:
  parameters:
    CustomerNumber:
      name: customerNumber
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  schemas:
    Address:
      type: object
      additionalProperties: false
      required:
        - line1
        - city
        - state
        - postalCode
        - countryCode
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        countryCode:
          type: string
    ArCustomer:
      type: object
      additionalProperties: false
      required:
        - customerNumber
        - customerName
        - customerAddress
        - customerCity
        - customerState
        - customerZipCode
        - customerPhone
      properties:
        customerNumber:
          type: integer
          format: int64
        customerName:
          type: string
        customerAddress:
          type: string
        customerCity:
          type: string
        customerState:
          type: string
        customerZipCode:
          type: string
        customerPhone:
          type: string
        activityCode:
          type: string
    ProCustomer:
      type: object
      additionalProperties: false
      required:
        - customerType
        - customerNumber
        - customerName
        - customerAddress
        - customerCity
        - customerState
        - customerZipCode
        - customerPhone
      properties:
        customerType:
          type: string
          const: pro
        customerNumber:
          type: integer
          format: int64
        customerName:
          type: string
        customerAddress:
          type: string
        customerCity:
          type: string
        customerState:
          type: string
        customerZipCode:
          type: string
        customerPhone:
          type: string
        accountNotes:
          type: string
        activityCode:
          type: string
        customerTypeDesc:
          type: string
    CashCustomer:
      type: object
      additionalProperties: false
      required:
        - customerType
        - customerNumbers
        - firstName
        - lastName
        - addresses
      properties:
        customerType:
          type: string
          const: cash
        customerNumbers:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
        firstName:
          type: string
        lastName:
          type: string
        preferredPhone:
          type: string
        preferredEmail:
          type: string
          format: email
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'